# applications/v4/api/openapi.yaml
openapi: 3.0.3
info:
  title: Weekly Planning API
  version: 1.0.0
  description: API contract for accounts, persons, projects, weeks, and weekly planning.
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []
paths:
...
  /persons:
    get:
      tags: [Persons]
      operationId: listPersons
      x-spring-paginated: true
      parameters:
        - $ref: '#/components/parameters/PageQueryParam'
        - $ref: '#/components/parameters/SizeQueryParam'
        - $ref: '#/components/parameters/SortQueryParam'
        - $ref: '#/components/parameters/SearchQueryParam'
        - $ref: '#/components/parameters/AccountRoleQueryParam'
        - $ref: '#/components/parameters/PersonStatusQueryParam'
      responses:
        '200': { $ref: '#/components/responses/PersonList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    post:
      tags: [Persons]
      operationId: createPerson
      requestBody:
        $ref: '#/components/requestBodies/CreatePersonRequestBody'
      responses:
        '201': { $ref: '#/components/responses/PersonCreated' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /persons/{personId}:
    get:
      tags: [Persons]
      operationId: getPersonById
      parameters:
        - $ref: '#/components/parameters/PersonIdPathParam'
      responses:
        '200': { $ref: '#/components/responses/PersonDetails' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    put:
      tags: [Persons]
      operationId: updatePersonById
      parameters:
        - $ref: '#/components/parameters/PersonIdPathParam'
      requestBody:
        $ref: '#/components/requestBodies/UpdatePersonRequestBody'
      responses:
        '200': { $ref: '#/components/responses/PersonDetails' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    delete:
      tags: [Persons]
      operationId: deletePersonById
      parameters:
        - $ref: '#/components/parameters/PersonIdPathParam'
      responses:
        '204': { $ref: '#/components/responses/NoContent' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalServerError' }
...
  /weeks:
    get:
      tags: [Weeks]
      operationId: listWeeks
      x-spring-paginated: true
      parameters:
        - $ref: '#/components/parameters/PageQueryParam'
        - $ref: '#/components/parameters/SizeQueryParam'
        - $ref: '#/components/parameters/SortQueryParam'
        - $ref: '#/components/parameters/WeekStatusQueryParam'
        - $ref: '#/components/parameters/WeekStartFromQueryParam'
        - $ref: '#/components/parameters/WeekStartToQueryParam'
      responses:
        '200': { $ref: '#/components/responses/WeekList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    post:
      tags: [Weeks]
      operationId: createWeek
      requestBody:
        $ref: '#/components/requestBodies/CreateWeekRequestBody'
      responses:
        '201': { $ref: '#/components/responses/WeekCreated' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /weeks/{weekId}:
    get:
      tags: [Weeks]
      operationId: getWeekById
      parameters:
        - $ref: '#/components/parameters/WeekIdPathParam'
      responses:
        '200': { $ref: '#/components/responses/WeekDetails' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    put:
      tags: [Weeks]
      operationId: updateWeekById
      description: ADMIN-only. When a week is set to COMPLETED, planning write operations are rejected.
      parameters:
        - $ref: '#/components/parameters/WeekIdPathParam'
      requestBody:
        $ref: '#/components/requestBodies/UpdateWeekRequestBody'
      responses:
        '200': { $ref: '#/components/responses/WeekDetails' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    delete:
      tags: [Weeks]
      operationId: deleteWeekById
      parameters:
        - $ref: '#/components/parameters/WeekIdPathParam'
      responses:
        '204': { $ref: '#/components/responses/NoContent' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalServerError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  headers:
    LocationHeader:
      description: URI of the created resource.
      schema:
        type: string
        format: uri
  parameters:
    PersonIdPathParam:
      name: personId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    WeekIdPathParam:
      name: weekId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    PageQueryParam:
      name: page
      in: query
      schema: { type: integer, format: int32, minimum: 0, default: 0 }
    SizeQueryParam:
      name: size
      in: query
      schema: { type: integer, format: int32, minimum: 1, maximum: 100, default: 20 }
    SortQueryParam:
      name: sort
      in: query
      schema:
        type: array
        items: { type: string }
      style: form
      explode: true
    SearchQueryParam:
      name: search
      in: query
      schema: { type: string, minLength: 1, maxLength: 120 }
    AccountRoleQueryParam:
      name: role
      in: query
      schema: { $ref: '#/components/schemas/AccountRole' }
    PersonStatusQueryParam:
      name: status
      in: query
      schema: { $ref: '#/components/schemas/PersonStatus' }
...
    WeekStatusQueryParam:
      name: status
      in: query
      schema: { $ref: '#/components/schemas/WeekStatus' }
    WeekStartFromQueryParam:
      name: weekStartFrom
      in: query
      schema: { type: string, format: date }
    WeekStartToQueryParam:
      name: weekStartTo
      in: query
      schema: { type: string, format: date }
...
  requestBodies:
...
    CreatePersonRequestBody:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/CreatePersonRequest' }
    UpdatePersonRequestBody:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/UpdatePersonRequest' }
...
    CreateWeekRequestBody:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/CreateWeekRequest' }
    UpdateWeekRequestBody:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/UpdateWeekRequest' }
...
  responses:
    NoContent:
      description: No content.
...
    PersonList:
      description: Paginated persons.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/PersonPage' }
    PersonDetails:
      description: Person details.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/PersonDetail' }
    PersonCreated:
      description: Person created.
      headers:
        Location: { $ref: '#/components/headers/LocationHeader' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/PersonDetail' }
...
    WeekList:
      description: Paginated weeks.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/WeekPage' }
    WeekDetails:
      description: Week details.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/WeekDetail' }
    WeekCreated:
      description: Week created.
      headers:
        Location: { $ref: '#/components/headers/LocationHeader' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/WeekDetail' }
...
    BadRequest:
      description: Bad request.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Unauthorized.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Forbidden.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: Conflict.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    ValidationError:
      description: Validation error.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    InternalServerError:
      description: Internal server error.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    AccountRole:
      type: string
      enum: [ADMIN, VIEWER]
    PersonStatus:
      type: string
      enum: [ACTIVE, INACTIVE]
...
    WeekStatus:
      type: string
      enum: [PLANNED, COMPLETED]
...
    SortDirection:
      type: string
      enum: [asc, desc]
...
    CreatePersonRequest:
      type: object
      required: [name, email, password, role, status]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        username:
          type: string
          minLength: 1
          maxLength: 128
        email:
          type: string
          format: email
          maxLength: 320
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 128
        role:
          $ref: '#/components/schemas/AccountRole'
        status:
          $ref: '#/components/schemas/PersonStatus'
    UpdatePersonRequest:
      type: object
      required: [name, email, role, status]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        username:
          type: string
          minLength: 1
          maxLength: 128
        email:
          type: string
          format: email
          maxLength: 320
        role:
          $ref: '#/components/schemas/AccountRole'
        status:
          $ref: '#/components/schemas/PersonStatus'
    PersonSummary:
      type: object
      required: [personId, accountId, username, name, email, role, status]
      properties:
        personId:
          type: string
          format: uuid
        accountId:
          type: string
          format: uuid
        username:
          type: string
          maxLength: 128
        name:
          type: string
          maxLength: 120
        email:
          type: string
          format: email
          maxLength: 320
        role:
          $ref: '#/components/schemas/AccountRole'
        status:
          $ref: '#/components/schemas/PersonStatus'
    PersonDetail:
      allOf:
        - $ref: '#/components/schemas/PersonSummary'
        - type: object
          required: [createdAt, updatedAt]
          properties:
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
    PersonPage:
      type: object
      required: [content, page]
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/PersonSummary'
        page:
          $ref: '#/components/schemas/PageMetadata'
...
    CreateWeekRequest:
      type: object
      required: [weekStart, weekEnd]
      properties:
        weekStart:
          type: string
          format: date
        weekEnd:
          type: string
          format: date
        status:
          allOf:
            - $ref: '#/components/schemas/WeekStatus'
          default: PLANNED
    UpdateWeekRequest:
      type: object
      required: [weekStart, weekEnd, status]
      properties:
        weekStart:
          type: string
          format: date
        weekEnd:
          type: string
          format: date
        status:
          $ref: '#/components/schemas/WeekStatus'
    WeekSummary:
      type: object
      required: [weekId, weekStart, weekEnd, status]
      properties:
        weekId:
          type: string
          format: uuid
        weekStart:
          type: string
          format: date
        weekEnd:
          type: string
          format: date
        status:
          $ref: '#/components/schemas/WeekStatus'
    WeekDetail:
      allOf:
        - $ref: '#/components/schemas/WeekSummary'
        - type: object
          required: [createdAt, updatedAt]
          properties:
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
    WeekPage:
      type: object
      required: [content, page]
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/WeekSummary'
        page:
          $ref: '#/components/schemas/PageMetadata'
...
    PageMetadata:
      type: object
      required: [page, size, totalElements, totalPages]
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        totalElements:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          format: int32
          minimum: 0
        sort:
          type: array
          items:
            $ref: '#/components/schemas/SortOrder'
    SortOrder:
      type: object
      required: [property, direction]
      properties:
        property:
          type: string
        direction:
          $ref: '#/components/schemas/SortDirection'
    FieldError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string
        rejectedValue:
          type: string
          nullable: true
    ErrorResponse:
      type: object
      required: [timestamp, status, error, code, message, path]
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        error:
          type: string
        code:
          type: string
        message:
          type: string
        path:
          type: string
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'